+++ modules/react-focus-navigation/src/__tests__/useActiveEventMap.spec.lua
@@ -17,7 +17,7 @@ local cleanup = ReactTestingLibrary.cleanup
 local Collections = require(Packages.Dev.Collections)
 local Object = Collections.Object
 
-local waitForEvents = require(script.Parent.waitForEvents)
+local waitForEvents = require(Packages.Utils).waitForEvents
 
 local FocusNavigationContext = require(script.Parent.Parent.FocusNavigationContext)
 local useActiveEventMap = require(script.Parent.Parent.useActiveEventMap)