+++ modules/react-focus-navigation/src/__tests__/useEventHandler.spec.lua
@@ -22,7 +22,7 @@ local GamePad = Rhodium.VirtualInput.GamePad
 local Collections = require(Packages.Dev.Collections)
 local Object = Collections.Object
 
-local waitForEvents = require(script.Parent.waitForEvents)
+local waitForEvents = require(Packages.Utils).waitForEvents
 
 local FocusNavigationContext = require(script.Parent.Parent.FocusNavigationContext)
 local useEventHandler = require(script.Parent.Parent.useEventHandler)