+++ modules/utils/src/init.lua
@@ -3,6 +3,7 @@ local createSignal = require(script.createSignal)
 local shallowEqual = require(script.shallowEqual)
 local getAncestors = require(script.getAncestors)
 local mockableWarn = require(script.mockableWarn)
+local waitForEvents = require(script.waitForEvents)
 
 export type Signal<T> = createSignal.Signal<T>
 export type FireSignal<T> = createSignal.FireSignal<T>
@@ -13,4 +14,5 @@ return {
 	shallowEqual = shallowEqual,
 	getAncestors = getAncestors,
 	mockableWarn = mockableWarn,
+	waitForEvents = waitForEvents,
 }